move enums and choices into models
